+++ typetreeflow/__init__.py
@@ -1,3 +1,3 @@
 """TypeTreeFlow package."""
 
-__version__ = "2.2.5"
+__version__ = "2.2.6"